Optimize docker image and Move from Python 3.6 to 3.9
The python version is updated from 3.6 to 3.9 because version 3.6 is no longer officially maintained on dockerhub.

The previous image was built on the full python image which caused the final size to be bloates, this commit builds the image in layers, first in the full python image so it can build the pip packages and then copies what was installed with pip to a slim python image which considerably reduces the size of the final image.