Clarify purpose
* Not everyone is aware of the limitations of Telegram API and as such
users may be beguiled into thinking that one could download files from
any public channels using this daemon.
* The reality of the matter is that Telegram API only allows you to
download from a channel in which you have admin rights.
* The desired workflow instead would be that for public channels; you
would manually forward files to a channel for which you have admin
rights over in order for this daemon to be of any use.